+++ templates/croct/ui/component/announcement-bar/code/page/react/page.tsx
@@ -11,7 +11,7 @@ export default function Page() {
           theme="light"
           title="Announcement bar"
           ctaLabel="Edit this content"
-          ctaLink="%workspaceUrl%/slots/edit/%slotId%/%slotVersion%?utm_medium=cli&utm_source=template&utm_campaign=00000000.CO.DE.ui_componentutm_content=announcement_bar&utm_term=react"
+          ctaLink="%workspaceUrl%/slots/edit/%slotId%/%slotVersion%?utm_medium=cli&utm_source=template&utm_campaign=00000000.CO.DE.ui_component&utm_content=announcement_bar&utm_term=react"
           ctaTarget="_blank"
           src="#"
           fullScreen
