+++ templates/npm/swiper/template.json5
