+++ tsconfig.json
@@ -33,5 +33,8 @@
   ],
   "ts-node": {
     "require": ["tsconfig-paths/register"]
-  }
+  },
+  "files": [
+    "slots.d.ts"
+  ]
 }
